refactor: 기물 구현체 공통 이동 좌표 계산 로직 부모 클래스로 이동
