+++ src/main/java/domain/piece/Cannon.java
@@ -1,23 +1,16 @@
 package domain.piece;
 
-import domain.coordinate.Path;
 import domain.coordinate.Position;
 import domain.Side;
 import domain.rule.CannonRule;
-import domain.rule.Rule;
 import domain.strategy.LinearStrategy;
-import domain.strategy.Strategy;
 
 import java.util.List;
-import java.util.Map;
 
 public final class Cannon extends Piece {
 
-    private final Strategy strategy = new LinearStrategy();
-    private final Rule rule = new CannonRule();
-
     public Cannon(Side side) {
-        super(side);
+        super(side, new LinearStrategy(), new CannonRule());
     }
 
     @Override
@@ -33,12 +26,5 @@ public final class Cannon extends Piece {
     @Override
     public boolean isEmpty() {
         return false;
-    }
-
-    @Override
-    public List<Position> getPossibleMoves(Position start, Pieces pieces) {
-        List<Path> paths = strategy.getPaths(start, pieces.getTopology());
-        Map<Position, Piece> pathPieces = pieces.collectPieces(paths);
-        return rule.getPossiblePositions(getSide(), pathPieces, paths);
     }
 }
