+++ src/main/java/domain/piece/Pawn.java
@@ -1,25 +1,17 @@
 package domain.piece;
 
-import domain.coordinate.Path;
 import domain.coordinate.Position;
 import domain.Side;
-import domain.rule.Rule;
 import domain.rule.StepRule;
 import domain.strategy.ForwardStrategy;
 import domain.strategy.StepStrategy;
-import domain.strategy.Strategy;
 
 import java.util.List;
-import java.util.Map;
 
 public final class Pawn extends Piece {
 
-    private final Strategy strategy;
-    private final Rule rule = new StepRule();
-
     public Pawn(Side side) {
-        super(side);
-        this.strategy = new ForwardStrategy(new StepStrategy(), forward());
+        super(side, new ForwardStrategy(new StepStrategy(), side.getForward()), new StepRule());
     }
 
     @Override
@@ -35,12 +27,5 @@ public final class Pawn extends Piece {
     @Override
     public boolean isEmpty() {
         return false;
-    }
-
-    @Override
-    public List<Position> getPossibleMoves(Position start, Pieces pieces) {
-        List<Path> paths = strategy.getPaths(start, pieces.getTopology());
-        Map<Position, Piece> pathPieces = pieces.collectPieces(paths);
-        return rule.getPossiblePositions(getSide(), pathPieces, paths);
     }
 }
