+++ src/main/java/domain/piece/Piece.java
@@ -1,17 +1,25 @@
 package domain.piece;
 
 import domain.coordinate.Direction;
+import domain.coordinate.Path;
 import domain.coordinate.Position;
+import domain.rule.Rule;
+import domain.strategy.Strategy;
 import domain.Side;
 
 import java.util.List;
+import java.util.Map;
 
 public abstract class Piece {
 
     private final Side side;
+    private final Strategy strategy;
+    private final Rule rule;
 
-    public Piece(Side side) {
+    public Piece(Side side, Strategy strategy, Rule rule) {
         this.side = side;
+        this.strategy = strategy;
+        this.rule = rule;
     }
 
     public Side getSide() {
@@ -30,8 +38,10 @@ public abstract class Piece {
         return this.side == side;
     }
 
-    public Direction forward() {
-        return side.getForward();
+    public List<Position> getPossibleMoves(Position start, Pieces pieces) {
+        List<Path> paths = strategy.getPaths(start, pieces.getTopology());
+        Map<Position, Piece> pathPieces = pieces.collectPieces(paths);
+        return rule.getPossiblePositions(getSide(), pathPieces, paths);
     }
 
     public abstract PieceType getType();
@@ -39,6 +49,4 @@ public abstract class Piece {
     public abstract Piece withSide(Side side);
 
     public abstract boolean isEmpty();
-
-    public abstract List<Position> getPossibleMoves(Position start, Pieces pieces);
 }
